Create cv2 main script
Note: This is intended for use in a Touch Designer project. The onCook function may be specific to TD, and the operators I am writing to are TD specific. This does not operate in python or a terminal window directly. 
Currently set to reposition transforms 1/3 frames, or 10 frames per second. This can be increased to fire every frame without too much CPU or GPU usage. 
Note that the onCook function does not follow the projects framerate. 
I left the time delay functions in there in case they are needed, but they are #checked out and do not apply until the # is removed. 
Email me with any questions: calebhmeyer@protonmail.com

Haarcascades documention relevant to this script: 
https://docs.opencv.org/4.x/d1/de5/classcv_1_1CascadeClassifier.html#aaf8181cb63968136476ec4204ffca498