Format Text Script.
This script, on every frame, reads text in 1 columns of every row, and adds that text, or integer, into a string in two columns of every row
If the cells number value is 42, the output would read 'face1/pt42:u' and 'face1/pt42:v'

This can be used for a wide range of projects. 
Please enjoy its usage and I hope it brings your project to life.